+++ lib/features/journal/state/entry_controller.g.dart
@@ -50,7 +50,7 @@ final class EntryControllerProvider
   }
 }
 
-String _$entryControllerHash() => r'1b21ac9f8debec64e15f45d097ccc13fa0a965d2';
+String _$entryControllerHash() => r'9a596204392046fb0183e822006abd0d938fcf97';
 
 final class EntryControllerFamily extends $Family
     with